Add DNS01-TXT support and restructure documentation
  Code changes:
  - Rename config field `dns01` → `dns01_txt` in config.go and ca.json
  - Remove `metrics.enabled` field (always-on now)
  - Add config_test.go with unit tests for ExternalCAS config parsing
  - Slim down external_test.go (tests moved to config_test.go)
  - Add `make test` target for running externalcas tests with -race

  Docs changes:
  - Add new configuration.md page with full field reference and upstream CA URL table
  - Remove configuration section from install.md (now its own page)
  - Remove pre-built binary install method from install.md
  - Rewrite README and docs landing page to lead with DNS01-TXT as a supported mode alongside EAB, and add LetsEncrypt as a viable upstream CA
  - Update quickstart.md to cover both EAB and DNS01-TXT configuration paths
  - Add dns01-txt.png sequence diagram asset
  - Regenerate Hugo public/ output
