+++ docs/content/client.md
@@ -1,6 +1,6 @@
 +++
 title = 'ACME Clients'
-weight = 30
+weight = 40
 BookToC = true
 +++
