+++ docs/content/install.md
@@ -6,14 +6,13 @@ BookToC = true
 
 # Install
 
-Three methods are available. The install script is recommended for most deployments.
+There are a few methods for produciton installation. The install script is recommended for most deployments due to it's simplicity. 
 
 | Method | Best for |
 |--------|---------|
 | [Install script](#install-script-recommended) | Standard Linux servers, systemd environments |
-| [Pre-built binary](#pre-built-binary) | Environments where curl-pipe-to-shell is prohibited |
 | [Build from source](#build-from-source) | Development, or architectures not covered by releases |
-| [Docker](#docker) | Container-based deployments |
+| [Docker](#docker) | Standalone container-based deployments 
 
 ---
 
@@ -54,71 +53,7 @@ curl -fsSL https://raw.githubusercontent.com/esnet/acme-proxy/main/install.sh | 
 | `$DB_DIR/bbolt` | bbolt KV store for ACME account state |
 | `/etc/systemd/system/acme-proxy.service` | Systemd service unit |
 
-The service is **enabled but not started**. [Configure `ca.json`](#configuration) before starting.
-
----
-
-## Pre-built Binary
-
-Download the release binary directly from the [GitHub releases page](https://github.com/esnet/acme-proxy/releases), verify the checksum, and install manually.
-
-```sh
-VERSION=1.0.0   # replace with the current release
-
-# Download binary and checksum
-curl -fsSLO "https://github.com/esnet/acme-proxy/releases/download/v${VERSION}/step-ca_linux_amd64"
-curl -fsSLO "https://github.com/esnet/acme-proxy/releases/download/v${VERSION}/step-ca_linux_amd64.sha256"
-
-# Verify
-sha256sum -c step-ca_linux_amd64.sha256
-
-# Install
-sudo install -o root -g root -m 0755 step-ca_linux_amd64 /opt/acme-proxy/step-ca
-```
-
-> For arm64, substitute `amd64` with `arm64` in the filename.
-
-After placing the binary, create the config directory and set up `ca.json` manually (see [Configuration](#configuration)), then create a systemd service unit:
-
-```sh
-sudo mkdir -p /opt/acme-proxy/db
-
-sudo tee /etc/systemd/system/acme-proxy.service <<'EOF'
-[Unit]
-Description=ACME Proxy Server (step-ca)
-Documentation=https://github.com/esnet/acme-proxy
-After=network-online.target
-Wants=network-online.target
-StartLimitIntervalSec=60
-StartLimitBurst=3
-
-[Service]
-Type=simple
-User=acme-proxy
-Group=acme-proxy
-ExecStart=/opt/acme-proxy/step-ca /opt/acme-proxy/ca.json
-WorkingDirectory=/opt/acme-proxy
-Restart=on-failure
-RestartSec=5
-NoNewPrivileges=yes
-ProtectSystem=strict
-PrivateTmp=yes
-AmbientCapabilities=CAP_NET_BIND_SERVICE
-CapabilityBoundingSet=CAP_NET_BIND_SERVICE
-ReadWritePaths=/opt/acme-proxy
-StandardOutput=journal
-StandardError=journal
-SyslogIdentifier=acme-proxy
-
-[Install]
-WantedBy=multi-user.target
-EOF
-
-sudo useradd -r -s /sbin/nologin acme-proxy
-sudo chown -R acme-proxy:acme-proxy /opt/acme-proxy
-sudo systemctl daemon-reload
-sudo systemctl enable acme-proxy
-```
+The service is **enabled but not started**. [Configure](configuration.md) `ca.json` file before starting.
 
 ---
 
@@ -140,16 +75,17 @@ make
 The build produces a `step-ca` binary in the current directory. Copy it to your install location:
 
 ```sh
-sudo install -o root -g root -m 0755 step-ca /opt/acme-proxy/step-ca
+# configure `ca.json` file before starting step-ca
+./step-ca ca.json
 ```
 
-Then follow the [pre-built binary](#pre-built-binary) instructions to create the service unit.
+Use the [installer script](https://raw.githubusercontent.com/esnet/acme-proxy/main/install.sh) as a reference to complete the setup with systemd service unit, service account user, permissions etc.
 
 ---
 
 ## Install using Docker
 
-Before starting the container on a linux distro
+Run the following commands before starting the container on a linux host
 
 ```sh
 mkdir -p /opt/acme-proxy/db
@@ -188,89 +124,6 @@ services:
       - ./db:/opt/acme-proxy/db
     restart: unless-stopped
 ```
-
----
-
-## Configuration
-
-All install methods use the same `ca.json` configuration format. The install script creates a template — five fields require customization before the service can start.
-
-### Minimal required configuration
-
-```json
-{
-  "address": ":443",
-  "dnsNames": ["acmeproxy.example.com"],
-  "logger": {
-    "format": "json"
-  },
-  "db": {
-    "type": "bbolt",
-    "dataSource": "/opt/acme-proxy/db/bbolt"
-  },
-  "authority": {
-    "type": "externalcas",
-    "config": {
-      "ca_url": "https://acme.sectigo.com/v2/InCommonRSAOV",
-      "account_email": "certadmin@example.com",
-      "eab_kid": "your-eab-key-id",
-      "eab_hmac_key": "your-eab-hmac-key",
-      "certlifetime": 30,
-      "metrics": {
-        "enabled": true,
-        "port": 9234,
-        "dataSource": "/opt/acme-proxy/db/metrics"
-      }
-    },
-    "provisioners": [
-      {
-        "type": "ACME",
-        "name": "acme",
-        "claims": {
-          "enableSSHCA": false,
-          "disableRenewal": false,
-          "allowRenewalAfterExpiry": false,
-          "disableSmallstepExtensions": true
-        }
-      }
-    ],
-    "backdate": "1m0s"
-  },
-  "tls": {
-    "minVersion": 1.2,
-    "maxVersion": 1.3,
-    "renegotiation": false
-  },
-  "commonName": "acmeproxy.example.com"
-}
-```
-
-### Field reference
-
-| Field | Required | Description |
-|-------|----------|-------------|
-| `address` | Yes | Listen address. `:443` binds all interfaces on port 443. |
-| `dnsNames` | Yes | Hostname(s) that this proxy is reachable at. acme-proxy requests a TLS cert for itself using these names on first start. |
-| `authority.config.ca_url` | Yes | ACME directory URL of your upstream certificate authority. |
-| `authority.config.account_email` | Yes | Email registered with the upstream CA. |
-| `authority.config.eab_kid` | Yes | External Account Binding Key ID, obtained from your CA's account portal. |
-| `authority.config.eab_hmac_key` | Yes | External Account Binding HMAC key, obtained from your CA's account portal. |
-| `authority.config.certlifetime` | No | Request certificate with a max lifetime period if supported by upstream CA |
-| `authority.config.metrics.enabled` | No | Expose Prometheus metrics. Default: `true`. |
-| `authority.config.metrics.port` | No | Metrics port. Default: `9234`. |
-| `db.dataSource` | Yes | Path to the bbolt KV store directory. Must be writable by the service user. |
-| `commonName` | Yes | Common name for the proxy's own TLS certificate. Should match `dnsNames[0]`. |
-
-### Upstream CA URLs
-
-| CA | ACME URL |
-|----|----------|
-| Sectigo / InCommon RSA OV | `https://acme.sectigo.com/v2/InCommonRSAOV` |
-| ZeroSSL | `https://acme.zerossl.com/v2/DV90` |
-
->[!IMPORTANT]
->**Note**<br>
-LetsEncrypt as a public certificate authority does not support ACME accounts via External Account Binding and hence it cannot be used as an upstream CA with acme-proxy.
 
 ---
