+++ docs/public/client/index.html
@@ -15,22 +15,22 @@ Table of Contents# Installing ACME Clients Account Registration Configuring Auto
 <meta name="theme-color" media="(prefers-color-scheme: light)" content="#ffffff">
 <meta name="theme-color" media="(prefers-color-scheme: dark)" content="#2e3440">
 <meta name="color-scheme" content="light dark"><meta property="og:url" content="http://localhost:1313/client/">
-  <meta property="og:site_name" content="acme-proxy">
-  <meta property="og:title" content="ACME Clients">
-  <meta property="og:description" content="ACME Clients# This guide covers installation and system-level configuration of ACME clients for use with acme-proxy. It is intended for system administrators deploying certificate automation on behalf of end users.
+	<meta property="og:site_name" content="acme-proxy">
+	<meta property="og:title" content="ACME Clients">
+	<meta property="og:description" content="ACME Clients# This guide covers installation and system-level configuration of ACME clients for use with acme-proxy. It is intended for system administrators deploying certificate automation on behalf of end users.
 For certificate issuance commands and per-scenario usage, see user.md.
 Table of Contents# Installing ACME Clients Account Registration Configuring Auto-Renewal via Systemd Log Management Installing ACME Clients# Certbot Note: Certbot’s actively maintained distribution is via Snap. The .deb packages available in apt repositories are no longer maintained by the Certbot project and ship outdated versions.">
-  <meta property="og:locale" content="en">
-  <meta property="og:type" content="article">
-    <meta property="article:modified_time" content="2026-05-09T16:30:55-05:00">
+	<meta property="og:locale" content="en">
+	<meta property="og:type" content="article">
+		<meta property="article:modified_time" content="2026-05-10T14:39:06-05:00">
 
 
-  <meta itemprop="name" content="ACME Clients">
-  <meta itemprop="description" content="ACME Clients# This guide covers installation and system-level configuration of ACME clients for use with acme-proxy. It is intended for system administrators deploying certificate automation on behalf of end users.
+	<meta itemprop="name" content="ACME Clients">
+	<meta itemprop="description" content="ACME Clients# This guide covers installation and system-level configuration of ACME clients for use with acme-proxy. It is intended for system administrators deploying certificate automation on behalf of end users.
 For certificate issuance commands and per-scenario usage, see user.md.
 Table of Contents# Installing ACME Clients Account Registration Configuring Auto-Renewal via Systemd Log Management Installing ACME Clients# Certbot Note: Certbot’s actively maintained distribution is via Snap. The .deb packages available in apt repositories are no longer maintained by the Certbot project and ship outdated versions.">
-  <meta itemprop="dateModified" content="2026-05-09T16:30:55-05:00">
-  <meta itemprop="wordCount" content="844">
+	<meta itemprop="dateModified" content="2026-05-10T14:39:06-05:00">
+	<meta itemprop="wordCount" content="844">
 
 <title>ACME Clients | acme-proxy</title>
 <link rel="icon" href="/favicon.png" >
@@ -40,7 +40,7 @@ Table of Contents# Installing ACME Clients Account Registration Configuring Auto
 
 
   <script defer src="/fuse.min.js"></script>
-  <script defer src="/en.search.min.3b53a73b50c77a03ef9f852162dd865a1ef44c853558a2c64c8f675e48958950.js" integrity="sha256-O1OnO1DHegPvn4UhYt2GWh70TIU1WKLGTI9nXkiViVA=" crossorigin="anonymous"></script>
+  <script defer src="/en.search.min.341d576d9d0fa3fc0e64d9ac46d2637f86d80f21c409734fc4ab6615fde12d8d.js" integrity="sha256-NB1XbZ0Po/wOZNmsRtJjf4bYDyHECXNPxKtmFf3hLY0=" crossorigin="anonymous"></script>
 
   <script defer src="/sw.min.6f6f90fcb8eb1c49ec389838e6b801d0de19430b8e516902f8d75c3c8bd98739.js" integrity="sha256-b2&#43;Q/LjrHEnsOJg45rgB0N4ZQwuOUWkC&#43;NdcPIvZhzk=" crossorigin="anonymous"></script>
 
@@ -141,6 +141,20 @@ Table of Contents# Installing ACME Clients Account Registration Configuring Auto
   
     <a href="/install/" class="">
       Install</a>
+  
+
+        </li>
+      
+    
+      
+        <li>
+          
+  
+  
+
+  
+    <a href="/configuration/" class="">
+      Configure</a>
   
 
         </li>
@@ -436,26 +450,26 @@ Table of Contents# Installing ACME Clients Account Registration Configuring Auto
 <p>All service units above write to the systemd journal with a unique <code>SyslogIdentifier</code>. Logs are accessible via <code>journalctl</code> and forwarded to syslog if your system runs rsyslog or syslog-ng with journal forwarding enabled.</p>
 <p><strong>Filter renewal logs by client:</strong></p>
 <table>
-  <thead>
-      <tr>
-          <th>Client</th>
-          <th>Command</th>
-      </tr>
-  </thead>
-  <tbody>
-      <tr>
-          <td>Certbot</td>
-          <td><code>journalctl -t certbot-renewal</code></td>
-      </tr>
-      <tr>
-          <td>acme.sh</td>
-          <td><code>journalctl -t acme-renewal</code></td>
-      </tr>
-      <tr>
-          <td>Lego</td>
-          <td><code>journalctl -t lego-renewal</code></td>
-      </tr>
-  </tbody>
+	<thead>
+			<tr>
+					<th>Client</th>
+					<th>Command</th>
+			</tr>
+	</thead>
+	<tbody>
+			<tr>
+					<td>Certbot</td>
+					<td><code>journalctl -t certbot-renewal</code></td>
+			</tr>
+			<tr>
+					<td>acme.sh</td>
+					<td><code>journalctl -t acme-renewal</code></td>
+			</tr>
+			<tr>
+					<td>Lego</td>
+					<td><code>journalctl -t lego-renewal</code></td>
+			</tr>
+	</tbody>
 </table>
 <p><strong>Follow logs in real time:</strong></p>
 <div class="highlight"><pre tabindex="0" style="color:#e2e4e5;background-color:#282a36;-moz-tab-size:4;-o-tab-size:4;tab-size:4;-webkit-text-size-adjust:none;"><code class="language-bash" data-lang="bash"><span style="display:flex;"><span>journalctl -t acme-renewal -f</span></span></code></pre></div><p><strong>Forward to syslog — rsyslog:</strong></p>
@@ -476,9 +490,9 @@ Table of Contents# Installing ACME Clients Account Registration Configuring Auto
   <div class="flex flex-wrap justify-between">
 
 <div>
-<a class="flex align-center" href="https://github.com/esnet/acme-proxy/commit/3c01d2951d66782f2f5844ce2a052e01f59427b0" title='Last modified by Kapil Agrawal | May 9, 2026' target="_blank" rel="noopener">
+<a class="flex align-center" href="https://github.com/esnet/acme-proxy/commit/e9ba8a117c1497712fa90429aad5bcaf9d343fd5" title='Last modified by Kapil Agrawal | May 10, 2026' target="_blank" rel="noopener">
     <img src="/icons/calendar.svg" class="book-icon" alt="Calendar" />
-    <span>May 9, 2026</span>
+    <span>May 10, 2026</span>
   </a>
 
 </div>
@@ -504,9 +518,9 @@ Table of Contents# Installing ACME Clients Account Registration Configuring Auto
   <div class="flex flex-wrap justify-between">
     <span>
     
-      <a href="/install/" class="flex align-center">
+      <a href="/configuration/" class="flex align-center">
         <img src="/icons/backward.svg" class="book-icon" alt="Backward" />
-        <span>Install</span>
+        <span>Configure</span>
       </a>
     
     </span>
