+++ docs/public/tags/index.html
@@ -12,13 +12,13 @@ https://github.com/alex-shpak/hugo-book
 <meta name="theme-color" media="(prefers-color-scheme: light)" content="#ffffff">
 <meta name="theme-color" media="(prefers-color-scheme: dark)" content="#2e3440">
 <meta name="color-scheme" content="light dark"><meta property="og:url" content="http://localhost:1313/tags/">
-  <meta property="og:site_name" content="acme-proxy">
-  <meta property="og:title" content="Tags">
-  <meta property="og:locale" content="en">
-  <meta property="og:type" content="website">
+	<meta property="og:site_name" content="acme-proxy">
+	<meta property="og:title" content="Tags">
+	<meta property="og:locale" content="en">
+	<meta property="og:type" content="website">
 
 
-  <meta itemprop="name" content="Tags">
+	<meta itemprop="name" content="Tags">
 
 <title>Tags | acme-proxy</title>
 <link rel="icon" href="/favicon.png" >
@@ -28,7 +28,7 @@ https://github.com/alex-shpak/hugo-book
 
 
   <script defer src="/fuse.min.js"></script>
-  <script defer src="/en.search.min.3b53a73b50c77a03ef9f852162dd865a1ef44c853558a2c64c8f675e48958950.js" integrity="sha256-O1OnO1DHegPvn4UhYt2GWh70TIU1WKLGTI9nXkiViVA=" crossorigin="anonymous"></script>
+  <script defer src="/en.search.min.341d576d9d0fa3fc0e64d9ac46d2637f86d80f21c409734fc4ab6615fde12d8d.js" integrity="sha256-NB1XbZ0Po/wOZNmsRtJjf4bYDyHECXNPxKtmFf3hLY0=" crossorigin="anonymous"></script>
 
   <script defer src="/sw.min.6f6f90fcb8eb1c49ec389838e6b801d0de19430b8e516902f8d75c3c8bd98739.js" integrity="sha256-b2&#43;Q/LjrHEnsOJg45rgB0N4ZQwuOUWkC&#43;NdcPIvZhzk=" crossorigin="anonymous"></script>
 
@@ -130,6 +130,20 @@ https://github.com/alex-shpak/hugo-book
   
     <a href="/install/" class="">
       Install</a>
+  
+
+        </li>
+      
+    
+      
+        <li>
+          
+  
+  
+
+  
+    <a href="/configuration/" class="">
+      Configure</a>
   
 
         </li>
