+++ docs/public/tags/index.xml
@@ -1,11 +1,15 @@
 <?xml version="1.0" encoding="utf-8" standalone="yes"?>
 <rss version="2.0" xmlns:atom="http://www.w3.org/2005/Atom">
-  <channel>
-    <title>Tags on acme-proxy</title>
-    <link>http://localhost:1313/tags/</link>
-    <description>Recent content in Tags on acme-proxy</description>
-    <generator>Hugo</generator>
-    <language>en</language>
-    <atom:link href="http://localhost:1313/tags/index.xml" rel="self" type="application/rss+xml" />
-  </channel>
+	<channel>
+		<title>Tags on acme-proxy</title>
+		<link>http://localhost:1313/tags/</link>
+		<description>Recent content in Tags on acme-proxy</description>
+		<generator>Hugo</generator>
+		<language>en</language>
+		
+		
+		
+		
+			<atom:link href="http://localhost:1313/tags/index.xml" rel="self" type="application/rss+xml" />
+	</channel>
 </rss>
