+++ externalcas/config.go
@@ -8,44 +8,47 @@ import (
 )
 
 // Configuration bits for a DNS provider as per Lego docs
-type LegoConfig struct {
+type legoConfig struct {
 	Provider       string            `json:"provider"`
 	DnsServersList []string          `json:"dns_servers"`
 	Env_Vars       map[string]string `json:"env_vars"`
 }
 
-type Metrics struct {
+type metrics struct {
 	Enabled    bool   `json:"enabled,omitempty"`
 	Port       int    `json:"port,omitempty"`
 	DataSource string `json:"datasource,omitempty"`
 }
 
 // AcmeProxyConfig contains the configuration for connecting to an external ACME CA
-type AcmeProxyConfig struct {
+type acmeProxyConfig struct {
 	// ACME directory url of External CA (required)
 	CaURL string `json:"ca_url"`
 
 	// External Account Binding
-	Email   string `json:"account_email,omitempty"`
+	Email string `json:"account_email,omitempty"`
+
 	Kid     string `json:"eab_kid"`
 	HmacKey string `json:"eab_hmac_key"`
 
 	// Certificate lifetime in days (optional)
 	CertLifetime int `json:"certlifetime,omitempty"`
 
-	// Lego provider connection variables for DNS01
-	Lego LegoConfig `json:"dns01"`
+	// Lego provider connection variables for dns01 TXT challenge
+	Lego legoConfig `json:"dns01_txt"`
 
 	// Prometheus metrics endpoint (optional)
-	Metrics Metrics `json:"metrics"`
+	Metrics metrics `json:"metrics"`
 
 	// derived during Validate(); not marshaled
-	useEAB   bool
-	useDNS01 bool
+	useEAB         bool
+	useDNS01       bool
+	metricsEnabled bool
 }
 
-// Validate checks if the AcmeProxyConfig contains required fields and valid values
-func (c *AcmeProxyConfig) Validate() error {
+// Validate checks if the values provided in ca.json file contain required fields
+// and valid values after they are unmarshalled into `acmeProxyConfig`
+func (c *acmeProxyConfig) Validate() error {
 	if c.CaURL == "" {
 		return errors.New("ca_url is required")
 	}
@@ -57,29 +60,37 @@ func (c *AcmeProxyConfig) Validate() error {
 	}
 
 	if !c.useEAB && !c.useDNS01 {
-		return errors.New("Missing eab or dns01 config. Must configure one.\nRefer docs https://software.es.net/acme-proxy/install/#configuration")
+		return errors.New("Missing eab or dns01 config. Must configure one\nRefer docs https://software.es.net/acme-proxy/configuration")
 	}
 	if c.CertLifetime < 0 {
 		return errors.New("certlifetime cannot be negative")
 	}
-	if c.Metrics.Enabled && c.Metrics.DataSource == "" {
-		return errors.New("metrics.datasource is required when metrics is enabled")
+
+	// Consider Metrics enabled only when port & datasource both are defined
+	if c.Metrics.Port > 0 && c.Metrics.DataSource != "" {
+		c.metricsEnabled = true
+	}
+
+	if (c.Metrics.Port > 0 && c.Metrics.DataSource == "") || (c.Metrics.Port == 0 && c.Metrics.DataSource != "") {
+		return errors.New("Invalid metrics port or datasource.\nRefer docs https://software.es.net/acme-proxy/configuration")
 	}
 	return nil
 }
 
 // HTTPTimeout returns the timeout for HTTP client operations
-func (c *AcmeProxyConfig) HTTPTimeout() time.Duration {
+func (c *acmeProxyConfig) HTTPTimeout() time.Duration {
 	return 90 * time.Second
 }
 
 // RequestTimeout returns the timeout for certificate request operations
-func (c *AcmeProxyConfig) RequestTimeout() time.Duration {
+func (c *acmeProxyConfig) RequestTimeout() time.Duration {
 	return 2 * time.Minute
 }
 
-func parseConfig(raw json.RawMessage) (*AcmeProxyConfig, error) {
-	var cfg AcmeProxyConfig
+// parseConfig is a helper function which reads ca.json file as rawjson and validates
+// acme-proxy specific configuration bits under the `authority` section of the config
+func parseConfig(raw json.RawMessage) (*acmeProxyConfig, error) {
+	var cfg acmeProxyConfig
 	if err := json.Unmarshal(raw, &cfg); err != nil {
 		return nil, fmt.Errorf("failed to unmarshal config: %w", err)
 	}
